auth-backend: move passport helpers into separate lib
