+++ plugins/auth-backend/src/lib/passport/PassportStrategyHelper.test.ts
