+++ plugins/auth-backend/src/providers/auth0/provider.ts
@@ -32,9 +32,8 @@ import {
   executeRefreshTokenStrategy,
   makeProfileInfo,
   PassportDoneCallback,
-  RedirectInfo,
-} from '../../lib/PassportStrategyHelper';
-import { AuthProviderConfig } from '../types';
+} from '../../lib/passport';
+import { AuthProviderConfig, RedirectInfo } from '../types';
 import { Config } from '@backstage/config';
 
 type PrivateInfo = {
