+++ plugins/auth-backend/src/providers/google/provider.ts
@@ -23,9 +23,8 @@ import {
   makeProfileInfo,
   executeFetchUserProfileStrategy,
   PassportDoneCallback,
-  RedirectInfo,
-} from '../../lib/PassportStrategyHelper';
-import { AuthProviderConfig } from '../types';
+} from '../../lib/passport';
+import { AuthProviderConfig, RedirectInfo } from '../types';
 import {
   OAuthProvider,
   OAuthProviderHandlers,
